+++ src/pages/Auth/index.jsx
@@ -1,0 +1,2 @@
+// SignIn SignUp
+
