+++ src/pages/Boards/_id.jsx
@@ -1,0 +1,21 @@
+// Details board
+import Container from '@mui/material/Container'
+import AppBar from '../../components/AppBar'
+import BoardBar from './BoardBar'
+import BoardContent from './BoardContent'
+
+const Board = () => {
+  return (
+    <>
+      <Container disableGutters={ true } maxWidth={ false } sx={{ height: '100vh' }}>
+        <AppBar />
+        <BoardBar />
+        <BoardContent />
+      </Container>
+    </>
+
+  )
+}
+
+export default Board
+
