+++ src/pages/Boards/index.jsx
@@ -1,0 +1,2 @@
+// Board list
+
