+++ src/pages/Users/_id.jsx
@@ -1,0 +1,2 @@
+// Users details
+
