+++ src/pages/Users/index.jsx
@@ -1,0 +1,2 @@
+// Users list
+
