+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+// Variables constants
+
